# finance-manager-backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This API serves personal finance endpoints. Accuracy is not garunteed",
        "title": "Go Finance Manager API",
        "contact": {},
        "version": "1.0.0"
    },
    "basePath": "/",
    "paths": {
...
        "/users/{userId}/credit-cards": {
            "get": {
                "description": "Returns an array of CreditCard objects belonging to a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Cards"
                ],
                "summary": "Get All User Credit Cards",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Search for Credit Cards by name",
                        "name": "search",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.CreditCard"
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Inserts a new Credit Card object for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Cards"
                ],
                "summary": "Insert Credit Card",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Credit card to insert",
                        "name": "creditCard",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.CreditCard"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
        "/users/{userId}/credit-cards/{ccId}": {
            "get": {
                "description": "Fetches a Credit Card by its ID for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Cards"
                ],
                "summary": "Get Credit Card by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the Credit Card",
                        "name": "ccId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.CreditCard"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a Credit Card by its ID for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Cards"
                ],
                "summary": "Update Credit Card by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the Credit Card",
                        "name": "ccId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a Credit Card by its ID for a given user",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Credit Cards"
                ],
                "summary": "Delete Credit Card by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "User ID",
                        "name": "userId",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "ID of the Credit Card",
                        "name": "ccId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/jsonutils.JSONResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "jsonutils.JSONResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.CreditCard": {
            "type": "object",
            "properties": {
                "apr": {
                    "type": "number"
                },
                "balance": {
                    "type": "number"
                },
                "createDt": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "lastUpdateDt": {
                    "type": "string"
                },
                "limit": {
                    "type": "number"
                },
                "minPayment": {
                    "type": "number"
                },
                "minPaymentPercentage": {
                    "type": "number"
                },
                "name": {
                    "type": "string"
                },
                "payment": {
                    "type": "number"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
    }
}